raspberry-pi/requirements.txt
# Dependências Python para Dashboard e Broker MQTT

# MQTT Cliente
paho-mqtt>=1.6.1

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# WebSocket para real-time
flask-socketio>=5.3.0

# Banco de Dados
sqlalchemy>=2.0.0

# Utilitários
python-dotenv>=1.0.0

